Scala Stack - Moving HTML and CSS to shared - Adding dependencies to compiler - Moving resources to script - Trying to use Scala JS version of Scala Tags but: [qtp315138752-20] INFO fiddle.Compiler - error: missing or invalid dependency detected while loading class file 'StyleSheet.class'. Could not access type Name in value sourcecode, because it (or its dependencies) are missing. Check your build definition for missing or conflicting dependencies. (Re-run with `-Ylog-classpath` to see the problematic classpath.) A full rebuild may help if 'StyleSheet.class' was compiled against an incompatible version of sourcecode.
[qtp315138752-20] INFO fiddle.Compiler - error: missing or invalid
dependency detected while loading class file 'StyleSheet.class'.
Could not access term sourcecode in package <root>,
because it (or its dependencies) are missing. Check your build
definition for
missing or conflicting dependencies. (Re-run with `-Ylog-classpath` to
see the problematic classpath.)
A full rebuild may help if 'StyleSheet.class' was compiled against an
incompatible version of <root>.
                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
[qtp315138752-20] INFO fiddle.Compiler - ScalaFiddle.scala:8: error:
could not find implicit value for parameter name: sourcecode.Name
  val hidden = cls(display := "none")
                  ^
t (or its dependencies) are missing. Check your build definition for
missing or conflicting dependencies. (Re-run with `-Ylog-classpath` to
see the problematic classpath.)
A full rebuild may help if 'StyleSheet.class' was compiled against an
incompatible version of <root>.
                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
2016-07-21 22:10:30.020:WARN:oejs.ServletHandler:qtp315138752-20:
/javascript.js
java.lang.Exception:
        at
fiddle.ScalaJsCompiler.compileScalaJsStrings(ScalaJsCompiler.scala:39)
        at
servlet.ScalaCompiler.doGet(JavaScriptCompilerServlet.scala:53)
        at javax.servlet.http.HttpServlet.service(HttpServlet.java:618)
        at javax.servlet.http.HttpServlet.service(HttpServlet.java:725)
        at
org.eclipse.jetty.servlet.ServletHolder.handle(ServletHolder.java:808)
        at
org.eclipse.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1669)
        at
com.googlecode.objectify.ObjectifyFilter.doFilter(ObjectifyFilter.java:48)
        at
org.eclipse.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1652)
        at
com.google.apphosting.utils.servlet.ParseBlobUploadFilter.doFilter(ParseBlobUploadFilter.java:125)
INFO     2016-07-21 22:10:30,033 module.py:788] default: "GET
/javascript.js HTTP/1.1" 500 303
        at
org.eclipse.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1652)
        at
com.google.apphosting.runtime.jetty9.SaveSessionFilter.doFilter(SaveSessionFilter.java:37)
        at
org.eclipse.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1652)
        at
com.google.apphosting.utils.servlet.TransactionCleanupFilter.doFilter(TransactionCleanupFilter.java:50)
        at
org.eclipse.jetty.servlet.ServletHandler$CachedChain.doFilter(ServletHandler.java:1652)
        at
org.eclipse.jetty.servlet.ServletHandler.doHandle(ServletHandler.java:585)
        at
org.eclipse.jetty.server.handler.ScopedHandler.handle(ScopedHandler.java:143)
        at
org.eclipse.jetty.security.SecurityHandler.handle(SecurityHandler.java:553)
        at
org.eclipse.jetty.server.session.SessionHandler.doHandle(SessionHandler.java:223)
        at
org.eclipse.jetty.server.handler.ContextHandler.doHandle(ContextHandler.java:1127)
        at
org.eclipse.jetty.servlet.ServletHandler.doScope(ServletHandler.java:515)
        at
org.eclipse.jetty.server.session.SessionHandler.doScope(SessionHandler.java:185)
        at
org.eclipse.jetty.server.handler.ContextHandler.doScope(ContextHandler.java:1061)
        at
com.google.apphosting.vmruntime.jetty9.VmRuntimeWebAppContext.doScope(VmRuntimeWebAppContext.java:299)
        at
org.eclipse.jetty.server.handler.ScopedHandler.handle(ScopedHandler.java:141)
        at
org.eclipse.jetty.server.handler.ContextHandlerCollection.handle(ContextHandlerCollection.java:215)
        at
org.eclipse.jetty.server.handler.HandlerCollection.handle(HandlerCollection.java:110)
        at
org.eclipse.jetty.server.handler.HandlerWrapper.handle(HandlerWrapper.java:97)
        at org.eclipse.jetty.server.Server.handle(Server.java:497)
        at
org.eclipse.jetty.server.HttpChannel.handle(HttpChannel.java:310)
        at
org.eclipse.jetty.server.HttpConnection.onFillable(HttpConnection.java:257)
        at
org.eclipse.jetty.io.AbstractConnection$2.run(AbstractConnection.java:540)
        at
org.eclipse.jetty.util.thread.QueuedThreadPool.runJob(QueuedThreadPool.java:635)
        at
org.eclipse.jetty.util.thread.QueuedThreadPool$3.run(QueuedThreadPool.java:555)
        at java.lang.Thread.run(Thread.java:745)
