+++ src/main/webapp/scalajs/script/ResourceValidator.scala
@@ -1,4 +1,4 @@
-package shared
+package script
 
 object ResourceValidator {
