+++ src/main/webapp/scalajs/shared/Css.scala
@@ -1,0 +1,9 @@
+package shared
+
+import scalatags.Text.all._
+import scalatags.stylesheet._
+
+trait Stylisch { this : StyleSheet =>
+
+  val hidden = cls(display := "none")
+}
